% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{PCA plot}
\usage{
pca(
  data = exp_data,
  center = T,
  retx = T,
  scale = FALSE,
  str_sample = "_.*m",
  str_group = "-.*"
)
}
\arguments{
\item{data}{iput data form the function 'pca_data_tidy'}

\item{center}{the prcomp param, detail see '?prcomp'}

\item{retx}{the prcomp param, detail see '?prcomp'}

\item{str_sample}{the 'regexp' for the sample name to become the target name}

\item{str_group}{the 'regexp' for the group name to become the target name}
}
\value{
a ggplot object
}
\description{
PCA plot
}
\examples{
pca(iris[,-5])
}
