% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_data_tidy.R
\name{pca_data_tidy}
\alias{pca_data_tidy}
\title{the Principal Components Analysis need the data must be transformed}
\usage{
pca_data_tidy(data)
}
\arguments{
\item{data}{the input data must be a dataframe or tibble which' rownames is gene id and colnames is sample name}
}
\value{
a dataframe
}
\description{
the Principal Components Analysis need the data must be transformed
}
\examples{
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")
pca_data_tidy(as.data.frame(test))
}
