% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{PCA plot}
\usage{
pca(
  data = data,
  center = T,
  retx = T,
  scale = FALSE,
  rename = c("diy", "replace"),
  sample_group = NULL,
  str_sample = NULL,
  str_group = "-.*"
)
}
\arguments{
\item{data}{iput data form the function 'pca_data_tidy'}

\item{center}{the prcomp param, detail see '?prcomp'}

\item{retx}{the prcomp param, detail see '?prcomp'}

\item{scale}{the prcomp param, detail see '?prcomp'}

\item{rename}{the method for change the sample and group names, two argment can choose, "diy" is for the creat a data for name,"replace" is use regexp to replace or change the name}

\item{sample_group}{a data for change the sample and group name, the rownames is sample and the first column is group}

\item{str_sample}{the 'regexp' for the sample name to become the target name}

\item{str_group}{the 'regexp' for the group name to become the target name}
}
\value{
a ggplot object
}
\description{
PCA plot
}
\examples{
pca(iris[,-5],sample_group = as.data.frame(iris[,5]))
}
